requirements.txt
discord.py
asyncio
pytz
discord-ext-bot